+++ movie/src/commonMain/kotlin/com/example/movie/data/remote/datasource/MovieDataSource.kt
@@ -1,11 +1,11 @@
 package com.example.movie.data.remote.datasource
 
+import com.example.core.data.model.movie.MovieResponse
 import com.example.core.data.model.movieCredits.MovieCreditsResponse
 import com.example.core.data.model.movieDetail.MovieDetailResponse
-import data.model.Movie.MovieResponse
 
 interface MovieDataSource {
-    suspend fun getDetails(id: Int): MovieDetailResponse
-    suspend fun getCredits(id: Int): MovieCreditsResponse
-    suspend fun getSimilar(id: Int): MovieResponse
+    suspend fun getDetails(id: Long): MovieDetailResponse
+    suspend fun getCredits(id: Long): MovieCreditsResponse
+    suspend fun getSimilar(id: Long): MovieResponse
 }
